+++ ios/MullvadVPN/View controllers/SelectLocation/DataSource/LocationNode.swift
@@ -21,6 +21,7 @@ class LocationNode: @unchecked Sendable {
     var isConnected: Bool
     var isSelected: Bool
     var isExcluded: Bool
+    let id: UUID = UUID()
 
     init(
         name: String,
@@ -46,6 +47,28 @@ class LocationNode: @unchecked Sendable {
         self.isConnected = isConnected
         self.isSelected = isSelected
         self.isExcluded = isExcluded
+    }
+
+    /// Recursively copies a node, its parent and its descendants from another
+    /// node (tree), with an optional custom root parent.
+    func copy(withParent parent: LocationNode? = nil) -> LocationNode {
+        let node = LocationNode(
+            name: name,
+            code: code,
+            locations: locations,
+            isActive: isActive,
+            parent: parent,
+            children: [],
+            showsChildren: showsChildren,
+            isHiddenFromSearch: isHiddenFromSearch,
+            isConnected: isConnected,
+            isSelected: false,  // explicity set to false since it's a different node
+            isExcluded: isExcluded
+        )
+
+        node.children = recursivelyCopyChildren(withParent: node)
+
+        return node
     }
 }
 
@@ -125,36 +148,11 @@ extension LocationNode {
                 }
             }
     }
-}
 
-extension LocationNode {
-    /// Recursively copies a node, its parent and its descendants from another
-    /// node (tree), with an optional custom root parent.
-    func copy(withParent parent: LocationNode? = nil) -> LocationNode {
-        let node = LocationNode(
-            name: name,
-            code: code,
-            locations: locations,
-            isActive: isActive,
-            parent: parent,
-            children: [],
-            showsChildren: showsChildren,
-            isHiddenFromSearch: isHiddenFromSearch,
-            isConnected: isConnected,
-            isSelected: false,  // explicity set to false since it's a different node
-            isExcluded: isExcluded
-        )
-
-        node.children = recursivelyCopyChildren(withParent: node)
-
-        return node
-    }
-
-    private func recursivelyCopyChildren(withParent parent: LocationNode) -> [LocationNode] {
+    fileprivate func recursivelyCopyChildren(withParent parent: LocationNode) -> [LocationNode] {
         children.map { $0.copy(withParent: parent) }
     }
 }
-
 extension LocationNode: Hashable {
     func hash(into hasher: inout Hasher) {
         hasher.combine(code)
@@ -214,4 +212,23 @@ class CustomListLocationNode: LocationNode, @unchecked Sendable {
             isHiddenFromSearch: isHiddenFromSearch
         )
     }
+
+    override func copy(withParent parent: LocationNode? = nil) -> CustomListLocationNode {
+        let node = CustomListLocationNode(
+            name: name,
+            code: code,
+            locations: locations,
+            isActive: isActive,
+            parent: parent,
+            children: [],
+            showsChildren: showsChildren,
+            isHiddenFromSearch: isHiddenFromSearch,
+            customList: customList
+        )
+
+        node.children = recursivelyCopyChildren(withParent: node)
+
+        return node
+    }
+
 }